+++ smartsheet-csharp-sdk/main/Smartsheet/Api/Internal/SightResourcesImpl.cs
@@ -50,11 +50,11 @@ namespace Smartsheet.Api.Internal
 
         /// <summary>
         /// <para>Gets the list of all Sights that the user has access to.</para>
-        /// 
+        ///
         /// <para>Mirrors to the following Smartsheet REST API method: GET /sights</para>
         /// </summary>
-        /// <returns>IndexResult object containing an array of Sight objects limited to the following attributes:
-        ///        id, name, accessLevel, permalink, createdAt, modifiedAt 
+        /// <returns>TokenPaginatedResult object containing an array of Sight objects limited to the following attributes:
+        ///        id, name, accessLevel, permalink, createdAt, modifiedAt
         /// </returns>
         /// <exception cref="System.InvalidOperationException"> if any argument is null or an empty string </exception>
         /// <exception cref="InvalidRequestException"> if there is any problem with the REST API request </exception>
@@ -62,33 +62,7 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="ResourceNotFoundException"> if the resource cannot be found </exception>
         /// <exception cref="ServiceUnavailableException"> if the REST API service is not available (possibly due to rate limiting) </exception>
         /// <exception cref="SmartsheetException"> if there is any other error during the operation </exception>
-        [Obsolete("This method is deprecated. Please use the overload that accepts TokenPaginationParameters instead.")]
-        public virtual PaginatedResult<Sight> ListSights(PaginationParameters? paging, DateTime? modifiedSince)
-        {
-            IDictionary<string, string> parameters = new Dictionary<string, string>();
-            if (paging != null)
-            {
-                parameters = paging.toDictionary();
-            }
-
-            return this.ListResourcesWithWrapper<Sight>("sights" + QueryUtil.GenerateUrl(null, parameters));
-        }
-
-        /// <summary>
-        /// <para>Gets the list of all Sights that the user has access to.</para>
-        /// 
-        /// <para>Mirrors to the following Smartsheet REST API method: GET /sights</para>
-        /// </summary>
-        /// <returns>IndexResult object containing an array of Sight objects limited to the following attributes:
-        ///        id, name, accessLevel, permalink, createdAt, modifiedAt 
-        /// </returns>
-        /// <exception cref="System.InvalidOperationException"> if any argument is null or an empty string </exception>
-        /// <exception cref="InvalidRequestException"> if there is any problem with the REST API request </exception>
-        /// <exception cref="AuthorizationException"> if there is any problem with the REST API authorization (access token) </exception>
-        /// <exception cref="ResourceNotFoundException"> if the resource cannot be found </exception>
-        /// <exception cref="ServiceUnavailableException"> if the REST API service is not available (possibly due to rate limiting) </exception>
-        /// <exception cref="SmartsheetException"> if there is any other error during the operation </exception>
-        public virtual TokenPaginatedResult<Sight> ListSights(TokenPaginationParameters? tokenPaging, DateTime? modifiedSince)
+        public virtual TokenPaginatedResult<Sight> ListSights(TokenPaginationParameters? tokenPaging = null)
         {
             IDictionary<string, string> parameters = new Dictionary<string, string>();
             if (tokenPaging != null)